Refactor IndicadorRepository and ValorIndicadorRepository to use Integer instead of Long for id field

+++ src/main/java/com/observatudo/backend/domain/repository/ValorIndicadorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.observatudo.backend.domain.model.ValorIndicador;
 
 @Repository
-public interface ValorIndicadorRepository extends JpaRepository<ValorIndicador, Integer> {
+public interface ValorIndicadorRepository extends JpaRepository<ValorIndicador, Long> {
     // Métodos de consulta personalizados podem ser definidos aqui, se necessário
 }
